tdf#46733 sw: don't double-select column during deleteCol
The problem here was that under IsNewModel tables were
selecting the entire column (including merged cells)
and then using THAT selection to ExpandColumnSelection.

Well, that ends in disaster. Since NewModel already
expands to catch the cells that are contained inside
the selected cells, there is no need to expand
the selection of GetTableCell.

Change-Id: I43089d912cdd17b2599619a43e40fecc2f5dd3b5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140455
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>