vcl: tdf#121327 PDF export for bitmap color fonts
This change extracts the PNG data from the sbix/CBDT tables and embeds
them as PDF Type 3 glyphs. In case the font supports both color layers
and color bitmaps, the color layer take priority.

This also reverts the part of the following commit that allowed bitmap
fonts when creating font subsets because this is not needed now as the
such fonts will now not reach the subsetting code.

commit 3e815dbedc1712ec376aa15d1d2bc522746804d5
Author: Khaled Hosny <khaledhosny@eglug.org>
Date:   Tue Aug 27 15:19:15 2019 +0200

    Make Noto Color Emoji font work on Linux

Change-Id: I350ec97956f37ae574956b22712869fd6d1a6990
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140457
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>