tdf#151143 Fix the size of Template Manager in ListView mode
A previous patch (709bb3e7ff161b98eada3cde53d8b3c0f18c78f9) fixed the size of the ThumbnailView in the Template Manager.

However it did not fix the size in ListView mode. So when the Template Manager opened in ListView, the height of the ListView the same as the number of available styles (rows in the list). The problem was most noticeable in Draw, which has just one template, but it actually affected all modules.

This patch fixes this issue in ListView.

Change-Id: Id6bf00b63e276b057ee191b191f57f295e9801cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140502
Tested-by: Jenkins
Tested-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>