vcl: accept bitmap fonts in OpenTTFontFile()
We never subset such fonts after:

commit 91a0ea040b345c27dbb1b49dd58ae722021c3e9b
Author: Khaled Hosny <khaled@aliftype.com>
Date:   Thu Sep 22 23:35:12 2022 +0200

    vcl: tdf#121327 PDF export for bitmap color fonts

But PrintFontManager::analyzeSfntFile() calls this function for
non-subsetting uses, it probably shouldn’t since all info it should
possibly need can come from FontConfig, but re-instate this for now.

This check is very blunt and it would lead to empty fonts in PDF if we
ever try to subset such a font, but lets clean this up later.

Change-Id: I804d0244ee24a4b4e92b36efd26b376cbe40784d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140542
Tested-by: Julien Nabet <serval2412@yahoo.fr>
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>