tdf#126457 - Add URL and workbook name to window titles
The Windows object allows to activate a window using its caption. The window caption is the text shown in the title bar at the top of the window. It can be changed which does not affect the name of the workbook. In MS these captions are unique, i.e., a file with the same file name stored in a different directory can't be open.

However, LibreOffice allows to open files with the same name. So in order to make it as closely to the functionality of MS, the windows activate method now works with URLs and changed window captions.

It has still to be clarified what happens if two workbooks have the same caption.

Change-Id: I77a8680bc64b5992191df2fa1a7d475f9a48f518
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140627
Tested-by: Jenkins
Reviewed-by: Andreas Heinisch <andreas.heinisch@yahoo.de>