tdf#151350: Fix extraneous gaps before marks
After latest changes we no longer need HarfBuzz buffer level
MONOTONE_CHARACTERS (which was needed to allow us to address individual
combining marks). With the default cluster level, combining marks get
the same cluster as their base and so we need not do any thing special.

I had to update testTdf139627 because this results in using ActualText
spans in the PDF, and PDFium support for them is lacking.

Change-Id: I0011072330fdbf409c30ff781fd3beaceae400f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140994
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>