tdf#150500 sw_fieldmarkhide: fix duplicate text frame on paste
The problem is that in MakeFrames() after the 1st iteration the
SwNode2Layout::NextFrame() on node 10 returns a SwTextFrame that was
just created by the 1st iteration, due to a fieldmark merging the 2
SwTextNodes.

Only itreate the frames that are pre-existing.

(regression from commit 0452c45fc3ee40243f144ab5f9af8a720c49db3b)

Change-Id: Iba684dfc903e1c72ca3e70a9aabf1aecf4baf32b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141032
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>