tdf#151315 Fix char style not applied if outline-folding is enabled
Making all folded content temporarily visible clobbers cursor
selections which is the cause of the character style not applied when
outline-folding is enabled. But it seems unnecessary to make all
folded content temporarily visible during style application. It would
have been nice if the author, me, had put a comment about why they
thought it was necessary. This patch drops making all folded content
temporarily visible during style application. If it is found to be
required, there are ways to include it without clearing cursor
selections.

Change-Id: Iedb1d0f350d5271a211300af083bc76eb06b1326
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141035
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>