tdf#145871 sw: ODF import: don't convert subtables if outer row ...
... has fixed or min height.

The code had 2 obvious problems: the fixed height on the outer row
wasn't cleared if the inner row didn't have a fixed height, and the code
to set lastSize on the last row erroneously set the first row's height
as well due to sharing the row format.

But it turns out that this doesn't work anyway in case any of the inner
rows are variable sized, because without layout it's not possible to
determine the height of these rows, and so the lastSize is going to be
too large in many cases.

(regression from commit 9485a94e70c9fe515a33469be4d06e270e9333fd)

Change-Id: I42ac7c14236562f9cae228efc0e98dc2fa8c2a23
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141079
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>