tdf#151060 sw PDF export: don't paint off-page part of drawing object
Reported to be a regression from
55338e167cbde45f0c4e426155c74cc0cff0b2d9 (DOCX import: fix lost objects
anchored to an empty linked header, 2020-01-10), the 3rd page of the PDF
export result contains an unexpected line shape.

This was "working" before as all objects anchored to the empty header
were lost.

Fix the problem by clipping the rendering to the page frame when
handling shapes, similar to what
c444113249df756c698e8b145eaf78259d5a0bab (tdf#91260 svx, sw: don't paint
off-page part of drawing object, 2016-12-06) did to fix the normal
rendering of the document.

The testcase document just has 2 pages, so there the unexpected shape
was on the 2nd page.

Change-Id: Ica24cd15717a1ee97dff448d385a10536671103e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141167
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>