add missing dependency for gbuild-to-json
while most of the integration targets are meant to have a nice debugging
environment, others like the vim-ide-integration or also the vscode one
seek to provide a nice editing experience with code completion and code
lookup.

Ideally you'd be able to have the integration available immediately and
can use the editor while a build is running in the background. But from
a fresh builddir make vim-ide-integration (vscode-ide-integration) fails
because of the missing dependency. That dependency probably is optional
for the purpose of just providing a editing environment as opposed to
providing build-targets like the integration does for some of the other
IDEs, so I consider this commit more of a "good enough" workaround for
the time being.

Change-Id: I586445fed849d416fde9804a1e2a9c6ee0ed50eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141248
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>