tdf#120326: properly prevent Windows from entering power saving
The previous call to SystemParametersInfoW with SPI_SETSCREENSAVEACTIVE
only prevented screen saver from running (which is not often used these
days), but screen was still turned off (and power saving was started).

Despite SetThreadExecutionState documentation [1] claims that it does
not stop the screen saver from executing, my experiments show that it
does (on my Windows 10). So just use it, because it is designed exactly
for this task (documentation discusses the presentation applications).

[1] https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-setthreadexecutionstate

Change-Id: I6eced8f0a7c17211a0b68851a40227eed9a22b1a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141289
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>