Fix turning off outline-folding doesn't clear all folded buttons
This happens, for unknown reason, when buttons that indicate outline
content is folded are near the current cursor position when the
options dialog is used to turn outline-folding off. This patch takes a
new approach to handle outline-folding options, that doesn't exhibit
this anomaly, by using the ShowOutlineContentVisibilityButton UNO
function. For the purpose of sanity and to aid understandability of
what is going on here, extra effort has been made to provide useful
comments and descriptive variable naming.

Change-Id: I8c3d4c207791f0ba151d799af22e1998aa7c3768
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141353
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>