tdf#151518 Do not tweak text area distances in Writer
The fix for tdf#148321 has introduced tweaks to the upper and lower
text distances for the case, that their sum is larger than the text
area height. That works well in import from PowerPoint but conflicts
with import of SmartArt shapes from Word.

Reason of the problem is not only the missing Twip to Hmm conversion in
TextBodyProperties::readjustTextDistances(). But I have not found which
detail in the tweak produces the wrong text location. For now I have
disabled the tweak for import from Word. Maybe someone finds a better
solution in future.

Because the user cannot insert such large distances in Word and SmartArt
shapes do not use such distances anyway, I think it does not harm to
disable the method for text documents.

Import from Excel cannot be tested because of tdf#83671.

Change-Id: I2e133f6864bba4419ad96e94d88aec86fe97d0fa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141414
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins