related tdf#151548 sw: use start of selection for GetCurrentFieldmark
When walking through a form, the current formfield is highlighted,
and pressing space should toggle a checkmarkField. Well, GetPoint
was pointing at the end of the selection, and so we were getting
a "sw/source/uibase/docvw/edtwin.cxx:2499: Where is my FieldMark??".

My attempt at a unit test caused random crashing (I assume it was
the protection-editing dialog box) and I never did succeed
in getting the checkbox to report an unchecked status anyway.

Change-Id: I0f28fb117a84cb5e56463a65d5bb72bbac174865
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141424
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins